Handle 'OK *' messages from imap server
My mail provider's imap server responds periodically with a "OK Still
here" message. As a result, a ValueError is raised when get_event tries
to unpack the message using space delim. This returns if the part does
not fit the format '<uid> <message>', instead of trying to take any
action on it.